Refactors code
We updated the auth endpoint to destructure the jwt token with pydantic. Additionally we are attempting to fix testing in CI by conditionally setting the minio connection. This commit also refactors earthly and docker compose to use profiles to spin up a dev instance and a prod instance for the application and fixes all earthly stages to work as intended